refactor: Normalise keyword `modify` -> `edit`.
