+++ helpers/roles/mod.ts
@@ -2,6 +2,6 @@ export * from "./addRole.ts";
 export * from "./createRole.ts";
 export * from "./deleteRole.ts";
 export * from "./editRole.ts";
+export * from "./editRolePositions.ts";
 export * from "./getRoles.ts";
-export * from "./modifyRolePositions.ts";
 export * from "./removeRole.ts";
